+++ service/arangod_config_builder.go
@@ -45,6 +45,19 @@ import (
 	"github.com/rs/zerolog"
 )
 
+var (
+	urlFixer = strings.NewReplacer(
+		"http://", "tcp://",
+		"https://", "ssl://",
+	)
+)
+
+// fixupEndpointURLSchemeForArangod changes endpoint URL schemes used by Starter to ones used by arangodb.
+// E.g. "http://localhost:8529" -> "tcp://localhost:8529"
+func fixupEndpointURLSchemeForArangod(u string) string {
+	return urlFixer.Replace(u)
+}
+
 // createArangodConf creates an arangod.conf file in the given host directory if it does not yet exists.
 // The arangod.conf file contains all settings that are considered static for the lifetime of the server.
 func createArangodConf(log zerolog.Logger, bsCfg BootstrapConfig, myHostDir, myContainerDir, myPort string, serverType ServerType, features DatabaseFeatures) ([]Volume, configFile, error) {
@@ -215,6 +228,13 @@ func createArangodArgs(log zerolog.Logger, config Config, clusterConfig ClusterC
 			optionPair{"--cluster.my-address", myTCPURL},
 			optionPair{"--cluster.my-role", "SINGLE"},
 		)
+	}
+	if serverType == ServerTypeCoordinator || serverType == ServerTypeResilientSingle {
+		if config.AdvertisedEndpoint != "" {
+			options = append(options,
+				optionPair{"--cluster.my-advertised-endpoint", fixupEndpointURLSchemeForArangod(config.AdvertisedEndpoint)},
+			)
+		}
 	}
 	if serverType != ServerTypeAgent && serverType != ServerTypeSingle {
 		for _, p := range clusterConfig.AllAgents() {